// src/store/enums.ts
// eslint-disable-next-line no-shadow
export enum MutationTypes {
  template_set = 'template_set',
  template_update = 'template_update',
  template_remove = 'template_remove',

  group_set = 'group_set',
  group_update = 'group_update',
  group_remove = 'group_remove',

  widget_set = 'widget_set',
  widget_update = 'widget_update',
  widget_remove = 'widget_remove',

  language_set = 'language_set',
  language_update = 'language_update',
  language_remove = 'language_remove',

  user_set = 'user_set',
  user_update = 'user_update',
  user_remove = 'user_remove',

  apiKey_set = 'apiKey_set',
  apiKey_update = 'apiKey_update',
  apiKey_remove = 'apiKey_remove',

  media_set = 'media_set',
  media_update = 'media_update',
  media_remove = 'media_remove',

  entryLite_set = 'entryLite_set',
  entryLite_update = 'entryLite_update',
  entryLite_remove = 'entryLite_remove',

  entry_set = 'entry_set',
  entry_update = 'entry_update',
  entry_remove = 'entry_remove',

  status_set = 'status_set',
  status_update = 'status_update',
  status_remove = 'status_remove',
}

// eslint-disable-next-line no-shadow
export enum GetterTypes {
  template_items = 'template_items',
  template_find = 'template_find',
  template_findOne = 'template_findOne',

  group_items = 'group_items',
  group_find = 'group_find',
  group_findOne = 'group_findOne',

  widget_items = 'widget_items',
  widget_find = 'widget_find',
  widget_findOne = 'widget_findOne',

  language_items = 'language_items',
  language_find = 'language_find',
  language_findOne = 'language_findOne',

  user_items = 'user_items',
  user_find = 'user_find',
  user_findOne = 'user_findOne',

  apiKey_items = 'apiKey_items',
  apiKey_find = 'apiKey_find',
  apiKey_findOne = 'apiKey_findOne',

  media_items = 'media_items',
  media_find = 'media_find',
  media_findOne = 'media_findOne',

  entryLite_items = 'entryLite_items',
  entryLite_find = 'entryLite_find',
  entryLite_findOne = 'entryLite_findOne',

  entry_items = 'entry_items',
  entry_find = 'entry_find',
  entry_findOne = 'entry_findOne',

  status_items = 'status_items',
  status_find = 'status_find',
  status_findOne = 'status_findOne',
}

// eslint-disable-next-line no-shadow
export enum ActionTypes {
  template_set = 'template_set',
  template_update = 'template_update',
  template_remove = 'template_remove',

  group_set = 'group_set',
  group_update = 'group_update',
  group_remove = 'group_remove',

  widget_set = 'widget_set',
  widget_update = 'widget_update',
  widget_remove = 'widget_remove',

  language_set = 'language_set',
  language_update = 'language_update',
  language_remove = 'language_remove',

  user_set = 'user_set',
  user_update = 'user_update',
  user_remove = 'user_remove',

  apiKey_set = 'apiKey_set',
  apiKey_update = 'apiKey_update',
  apiKey_remove = 'apiKey_remove',

  media_set = 'media_set',
  media_update = 'media_update',
  media_remove = 'media_remove',

  entryLite_set = 'entryLite_set',
  entryLite_update = 'entryLite_update',
  entryLite_remove = 'entryLite_remove',

  entry_set = 'entry_set',
  entry_update = 'entry_update',
  entry_remove = 'entry_remove',

  status_set = 'status_set',
  status_update = 'status_update',
  status_remove = 'status_remove',
}